refactor: make commands
